cast

+++ data_structure/sparse_table.hpp
@@ -19,15 +19,15 @@ template <class T> class sparse_table {
         table = {vec};
         log_table.clear();
 
-        for (int i = 0; (1 << i) < table[i].size(); i++) {
+        for (int i = 0; (1 << i) < (int)table[i].size(); i++) {
             table.push_back({});
-            for (int j = 0; j + (1 << i) < table[i].size(); j++) {
+            for (int j = 0; j + (1 << i) < (int)table[i].size(); j++) {
                 table[i + 1].push_back(op(table[i][j], table[i][j + (1 << i)]));
             }
         }
 
         log_table.resize(vec.size() + 1, 0);
-        for (int i = 2; i <= vec.size(); i++) {
+        for (int i = 2; i <= (int)vec.size(); i++) {
             log_table[i] = log_table[i >> 1] + 1;
         }
     }
